Add files via upload
here are the input functions, full to sparce matrix conversion function, and the output portion is in the projectScript file. I did notice that I missed the positions arrays, will make the changes on my files and update. Feel free to let me know if there are any other things to update